ai_translation_backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-docx==1.1.0
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.3
